+++ src/library/renderhud/ProfilerDebug.cpp
@@ -165,7 +165,7 @@ void ProfilerDebug::draw(uint64_t framecount, bool* p_open = nullptr)
 
         /* Convert frames to length here */
         const std::vector<TimeHolder>& frameTimings = Profiler::getFrameTimings();
-        if (frameTimings.size() < timeWindowFrames)
+        if (frameTimings.size() < static_cast<unsigned int>(timeWindowFrames))
             timeWindowFrames = frameTimings.size();
         
         if (timeWindowFrames > 0) {
@@ -223,7 +223,7 @@ void ProfilerDebug::draw(uint64_t framecount, bool* p_open = nullptr)
 
         const std::vector<TimeHolder>& frameTimings = Profiler::getFrameTimings();
         if (frameTimings.size() > 0) {
-            for (int f = 0; f < frameTimings.size() - 1; f++) {
+            for (long unsigned int f = 0; f < frameTimings.size() - 1; f++) {
                 /* Don't print frames that won't show up on screen */
                 if (frameTimings[f] < minTime)
                     continue;
